requirements.txt
requests==2.28.1
loguru==0.6.0
lxml==4.9.1
pyjwt
